load REPLs on demand
